+++ src/dotty/tools/dotc/core/SymbolLoaders.scala
@@ -13,7 +13,7 @@ import dotty.tools.io.{ ClassPath, AbstractFile }
 import Contexts._, Symbols._, Flags._, SymDenotations._, Types._, Scopes._, util.Positions._, Names._
 import StdNames._, NameOps._
 import Decorators.{StringDecorator, StringInterpolators}
-import pickling.ClassfileParser
+import classfile.ClassfileParser
 import scala.util.control.NonFatal
 
 object SymbolLoaders {
