+++ src/dotty/tools/dotc/core/tasty/TastyBuffer.scala
@@ -1,7 +1,7 @@
 package dotty.tools
 package dotc
 package core
-package pickling
+package tasty
 
 import util.Util.dble
