+++ src/dotty/tools/dotc/core/tasty/TastyReader.scala
@@ -1,8 +1,7 @@
 package dotty.tools
 package dotc
 package core
-package pickling
-
+package tasty
 
 import TastyBuffer._
 import TastyName.NameRef
